# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TLC platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/threads/{id}:
    get:
      operationId: getThread
      tags: [chat]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: oversight, in: query, required: false, schema: { type: boolean } }
      responses:
        "200":
          description: Thread detail (or null)
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/ThreadDetail" }
                  - { type: "null" }
  /chat/threads/{id}/messages:
    post:
      operationId: sendMessage
      tags: [chat]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageRequest" }
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendMessageResult" }
  /chat/threads/{id}/read:
    post:
      operationId: markThreadRead
      tags: [chat]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  # ----------------------------------------------------------- notifications
...
  /admin/companies:
    get:
      operationId: getCompaniesData
      tags: [admin]
      responses:
        "200":
          description: Companies data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompaniesData" }
    post:
      operationId: createCompany
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCompanyRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IdResult" }
  /admin/seats:
    post:
      operationId: purchaseSeats
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PurchaseSeatsRequest" }
      responses:
        "200":
          description: Purchased
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
  /admin/cohorts/clone:
    post:
      operationId: cloneCohort
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CloneCohortRequest" }
      responses:
        "200":
          description: Cloned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IdResult" }
...
components:
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    IdResult:
      type: object
      required: [ok, id]
      properties:
        ok: { type: boolean }
        id: { type: string }
...
    CompanyLite:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
...
    UserLite:
      type: object
      required: [id, email]
      properties:
        id: { type: string }
        name: { type: string, nullable: true }
        email: { type: string }
        image: { type: string, nullable: true }
        title: { type: string, nullable: true }
...
    MessageItem:
      type: object
      required: [id, body, senderId, createdAt]
      properties:
        id: { type: string }
        body: { type: string }
        senderId: { type: string }
        senderName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    ThreadMemberLite:
      type: object
      required: [userId, user]
      properties:
        userId: { type: string }
        user: { $ref: "#/components/schemas/UserLite" }
    ThreadDetail:
      type: object
      required: [id, type, members, messages]
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string, nullable: true }
        cohort:
          oneOf: [{ $ref: "#/components/schemas/CompanyLite" }, { type: "null" }]
        members:
          type: array
          items: { $ref: "#/components/schemas/ThreadMemberLite" }
        messages:
          type: array
          items: { $ref: "#/components/schemas/MessageItem" }
    SendMessageRequest:
      type: object
      required: [body]
      properties:
        body: { type: string }
    SendMessageResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        id: { type: string }
        error: { type: string }
...
    AdminCompanyRow:
      type: object
      required: [id, name, seatsAssigned, seatsTotal, viewers]
      properties:
        id: { type: string }
        name: { type: string }
        cohortName: { type: string, nullable: true }
        seatsAssigned: { type: integer }
        seatsTotal: { type: integer }
        viewers: { type: integer }
...
    CompaniesData:
      type: object
      required: [companies, cohorts]
      properties:
        companies:
          type: array
          items: { $ref: "#/components/schemas/AdminCompanyRow" }
        cohorts:
          type: array
          items: { $ref: "#/components/schemas/CompanyLite" }
    CreateCompanyRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        billingEmail: { type: string }
    PurchaseSeatsRequest:
      type: object
      required: [companyId, cohortId, quantity]
      properties:
        companyId: { type: string }
        cohortId: { type: string }
        quantity: { type: integer }
...
    CloneCohortRequest:
      type: object
      required: [sourceId, name, startDate]
      properties:
        sourceId: { type: string }
        name: { type: string }
        startDate: { type: string }
...